# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MobiManager API
  description: Complete API documentation for MobiManager - a mobile shop management SaaS platform
  version: 1.0.0
  contact:
    name: MobiManager Support
servers:
  - url: http://localhost:3000
    description: Local development server
security:
  - cookieAuth: []
paths:
...
  /api/admin/inventory/products:
    get:
      summary: List Products
      tags: [Inventory]
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: brandName
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: stockStatus
          in: query
          schema: { type: string, enum: [OUT_OF_STOCK, LOW_STOCK, IN_STOCK] }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: List of products
    post:
      summary: Create Product
      tags: [Inventory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProduct'
      responses:
        '201':
          description: Product created
...
  /api/admin/audit-logs:
    get:
      summary: List Audit Logs
      tags: [Audit Logs]
      parameters:
        - name: tableName
          in: query
          schema: { type: string }
        - name: recordId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: List of audit logs
...
components:
  schemas:
...
    CreateProduct:
      type: object
      required: [shopId, brandName, name, category, purchasePrice, sellingPrice, initialStock]
      properties:
        shopId: { type: string }
        brandName: { type: string }
        name: { type: string }
        category: { type: string, enum: [MOBILE, ACCESSORY] }
        accessoryType: { type: string }
        purchasePrice: { type: number }
        sellingPrice: { type: number }
        initialStock: { type: integer }
        lowStockAlertQty: { type: integer }
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: admin_token